part marking sets answered = true; fix update listeners
When the extension marks a part, it sets part.answered = true - it's now
up to the marking script to do this, since the marking algorithms
rewrite.

The update listeners, which set the part to dirty, were getting lost
because setBase64 doesn't happen synchronously. The eval_replacements
promise only resolves when setBase64 calls its callback.