big update
A lot of work to consolidate and document the code, allow links between
objects and part answers, and query the worksheet during part marking.

You can now create a blank worksheet, load a worksheet from a file
attached to the question, or from a geogebra.org material ID or base 64
string like before.

Object replacements can now be in the form of a dictionary setting a
variety of properties of the object in addition to its value.

The part links can now link GeoGebra objects to part answers, instead of
exercises to just extension parts. It looks like GeoGebra exercises have
been quietly dropped.

The link between objects and parts works both ways, so moving an object
in GeoGebra will update the input field, while changing the input field
will update the object in GeoGebra.

The `geogebra_applet` function now returns an object of type
`ggbapplet`, which can be queried in marking algorithms to get
properties of objects in the worksheet. These functions throw an error
if they're called before the applet has fully loaded.