TGGBApplet caches results from the applet
When resuming a question with a geogebra applet used in marking, the
marking algorithm will try to get values out of the geogebra applet, but
it has almost never finished loading when the marking algorithm runs.

This adds a cache to the TGGBApplet data type, storing all values
returned by the geogebra applet. They're included in the token's suspend
data, so when the question is resumed those values are available
immediately.