fix the size of the applet after loading
When loading a worksheet from a file, the wrapper element is given the
right size, but the inner panel is given height 0, for a reason I can't
work out.

Using the dimensions of the wrapper to resize the applet seems to work.
The applet seems to do some auto-resizing when the viewport is too small
anyway.